+++ src/Exceptions/Router/InvalidRouteException.php
@@ -1,0 +1,3 @@
+<?php
+namespace SimpleRoute\Exceptions\Router;
+class InvalidRouteException extends RouterException{}